chore: add variant system
